\name{Kira-package}

\alias{Kira-package}

\docType{package}

\title{Kira}

\description{Machine learning, containing several algorithms, in addition to functions that plot the graphs of the Receiver Operating Characteristic (ROC) and Precision-Recall (PRC) curve, and also a function that returns several metrics used to evaluate the models, the latter can be used in the classification results of other packages.}
 
\details{
\tabular{ll}{
Package: \tab Kira\cr
Type: \tab Package\cr
Version: \tab 1.0.0\cr
Date: \tab 2023-08-21\cr
License: \tab GPL(>= 2)\cr
LazyLoad: \tab yes\cr
}

This package contains:
\itemize{
  \item{}{Algorithms for supervised classification: knn, linear (lda) and quadratic (qda) discriminant analysis, linear regression, etc.}
  \item{}{Algorithms for unsupervised classification: hierarchical, kmeans, etc.}
  \item{}{A function that plots the ROC and PRC curve.}
  \item{}{A function that returns a series of metrics from models.}
}
}


\author{
Paulo Cesar Ossani <ossanipc@hotmail.com>
}

\references{

Aha, D. W.; Kibler, D.; Albert, M. K. Instance-based learning algorithms. \emph{Machine learning.} v.6, n.1, p.37-66. 1991.

Charnet, R.; at al.. \emph{Analise de modelos de regressao lienar,} 2a ed. Campinas: Editora da Unicamp, 2008. 357 p.

Chicco, D.; Warrens, M. J.; Jurman, G. The matthews correlation coefficient (mcc) is more informative than cohen's kappa and brier score in binary classification assessment. \emph{IEEE Access, IEEE}, v. 9, p. 78368-78381, 2021.

Ferreira, D. F. \emph{Estatistica Multivariada.} 2a ed. revisada e ampliada. Lavras: Editora UFLA, 2011. 676 p.

Mingoti, S. A. \emph{analysis de dados atraves de metodos de estatistica multivariada:} uma abordagem aplicada. Belo Horizonte: UFMG, 2005. 297 p.

Nicoletti, M. do C.. O modelo de aprendizado de maquina baseado em exemplares: principais caracteristicas e algoritmos. Sao Carlos: EdUFSCar, 2005. 61 p.

Rencher, A. C. \emph{Methods of multivariate analysis.} 2th. ed. New York: J.Wiley, 2002. 708 p.

Rencher, A. C.; Schaalje, G. B. \emph{Linear models in statisctic.} 2th. ed. New Jersey: John & Sons, 2008. 672 p.

Venabless, W. N. and Ripley, B. D. \emph{Modern Applied Statistics with S.} Fourth edition. Springer, 2002. 

}
